feat: improve help message
This PR will improve the help message for the command. It also moves the
flag definitions to a separate file, since the flag functionality is
likely to increase in complexity.
